refactor(OpenTelemetry): move to public API
