+++ telemetry/configuration_test.go
