+++ telemetry/counter.go
