+++ telemetry/counter_test.go
