+++ telemetry/counters_test.go
