+++ telemetry/interfaces.go
