+++ telemetry/interfaces_test.go
